Rewrote unitVector and vectorProjection
I rewrote unitVector to look more like the initial method, and I changed vectorProjection to how I thought it should look.  It didn't end up looking like yours, and this is probably a language thing (line 312).  To me, the vector which we are wanting to call the getIndexAtVector method upon is "v2.unitVector()", I believe before you had it written as "v2.unitVector()v2.getIndexAtVector(i).  Would like to know if my language is off.